+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## Unreleased
 
-- Add on-by-default crate feature `std`.
-- Add basic `Error` derive for enums.
+- Add basic `Error` derive for enums behind new on-by-default crate feature `std`.
+- Fix `#[no_std]` support.
 
 ## 0.1.3 - 2023-02-24