+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.1.4 - 2023-03-04
+
 - Add basic `Error` derive for enums.
 
 ## 0.1.3 - 2023-02-24